pyproject.toml
[tool.poetry]
name = "projct-practice"
version = "0.1.0"
description = ""
authors = ["limsu1015 <tngur1015@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.28.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
